rename(#269): ProfileImage -> Profile 폴더명 변경
