+++ src/app/coupon/[code]/page.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import Download from './_components/Download';
 import Button from '@/components/Button/Button';
 import DownloadCoupon from '@/components/Coupon/DownloadCoupon';
-import ProfileImage from '@/components/ProfileImage/ProfileImage';
+import ProfileImage from '@/components/Profile/ProfileImage';
 import { profileInfo } from '@/types/auth';
 import { IprivateCoupon } from '@/types/coupon';
