+++ src/components/ClassPreview/ResponsiveClassPreview.tsx
@@ -2,7 +2,7 @@
 import { useState } from 'react';
 import { StarSVG } from '@/icons/svg';
 import Carousel from '../Carousel/Carousel';
-import ProfileImage from '../ProfileImage/ProfileImage';
+import ProfileImage from '../Profile/ProfileImage';
 import { ClassCardType } from '@/types/class';
 import { useRouter } from 'next/navigation';
