+++ src/components/Profile/ProfileImage.stories.tsx
